move make_gemm_mapping to separate file
