+++ main_aie_codegen_gemm_col.py
@@ -3,7 +3,8 @@ import logging as _logging
 import re
 
 from stream.api import optimize_allocation_co
-from stream.inputs.aie.workload.make_onnx_gemm import make_gemm_mapping_single_col, make_gemm_workload
+from stream.inputs.aie.mapping.make_gemm_mapping import make_gemm_mapping_single_col
+from stream.inputs.aie.workload.make_onnx_gemm import make_gemm_workload
 
 _logging_level = _logging.INFO
 _logging_format = "%(asctime)s - %(name)s.%(funcName)s +%(lineno)s - %(levelname)s - %(message)s"
